Deep debug.

+++ spikeinterface/core/npzsortingextractor.py
@@ -52,7 +52,7 @@ class NpzSortingExtractor(BaseSorting):
             spike_labels = []
             for unit_id in units_ids:
                 sp_ind = sorting.get_unit_spike_train(unit_id, segment_index=seg_index)
-                spike_indexes.append(sp_ind)
+                spike_indexes.append(sp_ind.astype('int64'))
                 # spike_labels.append(np.ones(sp_ind.size, dtype='int64')*unit_id)
                 spike_labels.append(np.array([unit_id] * sp_ind.size))
 
@@ -66,7 +66,6 @@ class NpzSortingExtractor(BaseSorting):
             else:
                 spike_indexes = np.array([], dtype='int64')
                 spike_labels = np.array([], dtype='int64')
-
             d[f'spike_indexes_seg{seg_index}'] = spike_indexes
             d[f'spike_labels_seg{seg_index}'] = spike_labels
